+++ Archer/globals.h
@@ -1,5 +1,5 @@
 #pragma once
 
 const float m_gravity{ 0.035f };
-const float m_screenWidth{ 800.0f };
-const float m_screenHeight{ 600.0f };
+const unsigned int m_screenWidth{ 800u };
+const unsigned int m_screenHeight{ 600u };
